# api-doc/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CreateUserDto": {
				"description": "用户基础数据传输对象",
				"properties": {
					"userName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"description": "可选的邮箱地址"
					},
					"permissionId": {
						"type": "number",
						"format": "double",
						"description": "可选的权限ID"
					},
					"status": {
						"type": "boolean",
						"description": "用户状态标志"
					}
				},
				"required": [
					"userName",
					"phone",
					"password",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"description": "用户登录请求对象",
				"properties": {
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"phone",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_any.string-or-number-or-any_": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CreateMenuDto": {
				"properties": {
					"icon": {
						"type": "string"
					},
					"menuName": {
						"type": "string"
					},
					"menuPath": {
						"type": "string"
					},
					"permNode": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"permissionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"menuName",
					"menuPath",
					"level",
					"parentId",
					"permissionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SearchParams": {
				"description": "用户搜索参数对象",
				"properties": {
					"userName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"description": "可选的邮箱地址"
					},
					"permissionId": {
						"type": "number",
						"format": "double",
						"description": "可选的权限ID"
					},
					"status": {
						"type": "boolean",
						"description": "用户状态标志"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userName",
					"phone",
					"password",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateUserDto_": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"description": "可选的邮箱地址"
					},
					"permissionId": {
						"type": "number",
						"format": "double",
						"description": "可选的权限ID"
					},
					"status": {
						"type": "boolean",
						"description": "用户状态标志"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"PermissionDto": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"roleName": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"permissionScope": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"userName",
					"roleName",
					"members",
					"permissionScope"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_PermissionDto.Exclude_keyofPermissionDto.userName__": {
				"properties": {
					"roleName": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"permissionScope": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"roleName",
					"members",
					"permissionScope"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_PermissionDto.userName_": {
				"$ref": "#/components/schemas/Pick_PermissionDto.Exclude_keyofPermissionDto.userName__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_PermissionDto_": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"roleName": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"permissionScope": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tsoa-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LiuQi"
		}
	},
	"paths": {
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"permissionId": {
													"type": "number",
													"format": "double"
												},
												"status": {
													"type": "boolean"
												},
												"createDate": {
													"type": "string",
													"format": "date-time"
												},
												"email": {
													"type": "string"
												},
												"phone": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"userName": {
													"type": "string"
												},
												"userId": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"permissionId",
												"status",
												"createDate",
												"email",
												"phone",
												"password",
												"userName",
												"userId"
											],
											"type": "object"
										},
										"token": {}
									},
									"required": [
										"user",
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/users/logout": {
			"get": {
				"operationId": "LoginOut",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/info": {
			"get": {
				"operationId": "GetUserInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"menuList": {
													"items": {
														"$ref": "#/components/schemas/CreateMenuDto"
													},
													"type": "array"
												},
												"user": {
													"$ref": "#/components/schemas/Pick_any.string-or-number-or-any_"
												}
											},
											"required": [
												"menuList",
												"user"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										},
										"code": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"data",
										"message",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/list": {
			"post": {
				"operationId": "GetUserList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"permissionId": {
												"type": "number",
												"format": "double"
											},
											"status": {
												"type": "boolean"
											},
											"createDate": {
												"type": "string",
												"format": "date-time"
											},
											"email": {
												"type": "string"
											},
											"phone": {
												"type": "string"
											},
											"password": {
												"type": "string"
											},
											"userName": {
												"type": "string"
											},
											"userId": {
												"type": "number",
												"format": "double"
											},
											"permission": {
												"properties": {
													"roleName": {
														"type": "string"
													}
												},
												"required": [
													"roleName"
												],
												"type": "object"
											},
											"role": {
												"type": "string"
											}
										},
										"required": [
											"permissionId",
											"status",
											"createDate",
											"email",
											"phone",
											"password",
											"userName",
											"userId",
											"permission",
											"role"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchParams"
							}
						}
					}
				}
			}
		},
		"/users/delete": {
			"get": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/update": {
			"post": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Partial_CreateUserDto_"
									},
									{
										"properties": {
											"userId": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"userId"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
		"/users/read": {
			"get": {
				"operationId": "GetUserDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"permissionId": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "boolean"
										},
										"createDate": {
											"type": "string",
											"format": "date-time"
										},
										"email": {
											"type": "string"
										},
										"phone": {
											"type": "string"
										},
										"userName": {
											"type": "string"
										},
										"userId": {
											"type": "number",
											"format": "double"
										},
										"password": {}
									},
									"required": [
										"permissionId",
										"status",
										"createDate",
										"email",
										"phone",
										"userName",
										"userId",
										"password"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"用户接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/permission/list": {
			"post": {
				"operationId": "GetPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"roleName": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"roleName",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "获取权限列表",
				"tags": [
					"权限管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "查询参数",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PermissionDto",
								"description": "查询参数"
							}
						}
					}
				}
			}
		},
		"/permission/add": {
			"post": {
				"operationId": "AddPermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"roleName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"roleName",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "添加权限",
				"tags": [
					"权限管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "权限信息",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_PermissionDto.userName_",
								"description": "权限信息"
							}
						}
					}
				}
			}
		},
		"/permission/update": {
			"post": {
				"operationId": "UpdatePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"roleName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"roleName",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "更新权限",
				"tags": [
					"权限管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "权限更新数据",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Partial_PermissionDto_"
									},
									{
										"properties": {
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"id"
										],
										"type": "object"
									}
								],
								"description": "权限更新数据"
							}
						}
					}
				}
			}
		},
		"/permission/delete": {
			"get": {
				"operationId": "DeletePermission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"roleName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"roleName",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "删除权限",
				"tags": [
					"权限管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "权限ID",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}